Added test helper class
